+++ clairvoyance/__init__.py
@@ -1,0 +1,1 @@
+from clairvoyance.cli import cli  # noqa
