+++ clairvoyance/entities/__init__.py
@@ -1,1 +1,1 @@
-from clairvoyance.entities.primitives import *
+from clairvoyance.entities.primitives import *  # noqa
