+++ pyproject.toml
@@ -22,8 +22,14 @@ readme = "README.md"
 "repository" = "https://github.com/Escape-Technologies/clairvoyance"
 
 [tool.poetry.scripts]
-clairvoyance = 'clairvoyance:main'
+clairvoyance = 'clairvoyance:cli'
 
+[tool.coverage.run]
+omit = [
+    'tests/*',
+    'examples/*',
+    'docs/*',
+]
 
 [tool.poetry.dependencies]
 python = ">=3.7,<=3.11"
